fix(schedule): apply the node→bun shim + live env to the oxmgr path too
Review tracing found the same bug class on 'ay schedule': oxmgr's
'#!/usr/bin/env node' launcher is spawned without an explicit env and the
schedule path never established a node runtime at all, so a bun-only box
dies at the first oxmgr call. Extract ensureNodeRuntime/liveEnv into
ts/nodeRuntime.ts (now on the shared agentYesHome helper), call
ensureNodeRuntime at cmdSchedule entry, and pass liveEnv() in
schedule.run() and oxmgrService.ensureBootAutostart. Also tighten the
source-scan guard to require a real liveEnv() option rather than the
substring 'env:'.
